[refactor] move trafficLight class to its module
