+++ 1_The_Physical/Level_2_Stateful_Guess_Point/Shared/TrafficLight/src/trafficLight.spec.ts
@@ -1,87 +1,4 @@
-enum Status {
-  Red = "red",
-  Yellow = "yellow",
-  Green = "green",
-  Off = "off",
-  Boot = "boot",
-}
-
-type TrafficLightProps = {
-  readonly currentStatus: Status;
-};
-
-class TrafficLight {
-  private props: TrafficLightProps;
-
-  constructor() {
-    this.props = {
-      currentStatus: Status.Off,
-    };
-  }
-
-  isOff() {
-    return this.props.currentStatus === Status.Off;
-  }
-
-  isOn() {
-    return this.props.currentStatus !== Status.Off;
-  }
-
-  turnOn() {
-    if (this.isOff())
-      this.props = {
-        currentStatus: Status.Boot,
-      };
-  }
-
-  turnOff() {
-    if (this.isOn())
-      this.props = {
-        currentStatus: Status.Off,
-      };
-  }
-
-  advance() {
-    if (this.isBoot()) {
-      this.advanceTo(Status.Green);
-      return;
-    }
-    if (this.isGreen()) {
-      this.advanceTo(Status.Yellow);
-      return;
-    }
-    if (this.isYellow()) {
-      this.advanceTo(Status.Red);
-      return;
-    }
-    if (this.isRed()) {
-      this.advanceTo(Status.Green);
-      return;
-    }
-  }
-
-  isBoot() {
-    return this.props.currentStatus === Status.Boot;
-  }
-
-  isGreen() {
-    return this.props.currentStatus === Status.Green;
-  }
-
-  isYellow() {
-    return this.props.currentStatus === Status.Yellow;
-  }
-
-  isRed() {
-    return this.props.currentStatus === Status.Red;
-  }
-
-  private advanceTo(status: Status) {
-    this.props = {
-      currentStatus: status,
-    };
-  }
-}
+import { TrafficLight } from "./trafficLight";
 
 describe("TrafficLight", () => {
   it("should be defined", () => {
